refactor(descriptors): split GC root scanning into descriptor_state/gc_scan
