+++ crates/perry-runtime/src/object/prototype_helpers.rs
@@ -4,8 +4,7 @@ pub(crate) fn constructor_dynamic_prototype(obj: *const ObjectHeader) -> Option<
     if obj.is_null() {
         return None;
     }
-    let key =
-        crate::string::canonical_key(b"constructor");
+    let key = crate::string::canonical_key(b"constructor");
     let constructor = js_object_get_field_by_name_f64(obj, key);
     let bits = constructor.to_bits();
     let top16 = bits >> 48;
